requirements.txt
aniso8601==8.1.0
click==7.1.2
dill==0.3.3
Flask==1.1.2
Flask-MySQL==1.5.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
klepto==0.2.0

MarkupSafe==1.1.1
mpmath==1.1.0
MyApplication==0.1.0
mysql==0.0.2
mysqlclient==2.0.3
mystic==0.3.7
numpy==1.19.5
pox==0.2.9
PyMySQL==1.0.2
pytz==2020.5
six==1.15.0
sympy==1.7.1
Werkzeug==1.0.1

simplejson~=3.17.2